Modify edxlocal role to be able to create databases on RDS.
The database creation scripts in the edxlocal role only work if MySQL
is hosted locally. Letting ansible take care of creating the databases
is very convenient, so modify the script to be able to create
databases on remote servers.

Having separate users for each part of the edX stack has only been
added very recently, and we are ignoring it here...